MIGENG-67 SourceRampDownCosts rule and test (#5)
* MIGENG-67 SourceRampDownCosts rule and test

* MIGENG-67 Done some changes to rule, bean and test (#1)

+++ src/main/java/org/jboss/xavier/analytics/pojo/output/SourceRampDownCostsModel.java
@@ -3,45 +3,45 @@ package org.jboss.xavier.analytics.pojo.output;
 public class SourceRampDownCostsModel
 {
     // SourceRampDownCosts
-    private Double year1ServersOffSource;
-    private Double year1SourceActiveLicense;
-    private Double year1SourcePaidMaintenance;
+    private Integer year1ServersOffSource;
+    private Integer year1SourceActiveLicense;
+    private Integer year1SourcePaidMaintenance;
     private Double year1SourceMaintenancePerServerValue;
     private Double year1SourceMaintenanceTotalValue;
-    private Double year2ServersOffSource;
-    private Double year2SourceActiveLicense;
-    private Double year2SourcePaidMaintenance;
+    private Integer year2ServersOffSource;
+    private Integer year2SourceActiveLicense;
+    private Integer year2SourcePaidMaintenance;
     private Double year2SourceMaintenancePerServerValue;
     private Double year2SourceMaintenanceTotalValue;
-    private Double year3ServersOffSource;
-    private Double year3SourceActiveLicense;
-    private Double year3SourcePaidMaintenance;
+    private Integer year3ServersOffSource;
+    private Integer year3SourceActiveLicense;
+    private Integer year3SourcePaidMaintenance;
     private Double year3SourceMaintenancePerServerValue;
     private Double year3SourceMaintenanceTotalValue;
 
     public SourceRampDownCostsModel() {}
 
-    public Double getYear1ServersOffSource() {
+    public Integer getYear1ServersOffSource() {
         return year1ServersOffSource;
     }
 
-    public void setYear1ServersOffSource(Double year1ServersOffSource) {
+    public void setYear1ServersOffSource(Integer year1ServersOffSource) {
         this.year1ServersOffSource = year1ServersOffSource;
     }
 
-    public Double getYear1SourceActiveLicense() {
+    public Integer getYear1SourceActiveLicense() {
         return year1SourceActiveLicense;
     }
 
-    public void setYear1SourceActiveLicense(Double year1SourceActiveLicense) {
+    public void setYear1SourceActiveLicense(Integer year1SourceActiveLicense) {
         this.year1SourceActiveLicense = year1SourceActiveLicense;
     }
 
-    public Double getYear1SourcePaidMaintenance() {
+    public Integer getYear1SourcePaidMaintenance() {
         return year1SourcePaidMaintenance;
     }
 
-    public void setYear1SourcePaidMaintenance(Double year1SourcePaidMaintenance) {
+    public void setYear1SourcePaidMaintenance(Integer year1SourcePaidMaintenance) {
         this.year1SourcePaidMaintenance = year1SourcePaidMaintenance;
     }
 
@@ -61,27 +61,27 @@ public class SourceRampDownCostsModel
         this.year1SourceMaintenanceTotalValue = year1SourceMaintenanceTotalValue;
     }
 
-    public Double getYear2ServersOffSource() {
+    public Integer getYear2ServersOffSource() {
         return year2ServersOffSource;
     }
 
-    public void setYear2ServersOffSource(Double year2ServersOffSource) {
+    public void setYear2ServersOffSource(Integer year2ServersOffSource) {
         this.year2ServersOffSource = year2ServersOffSource;
     }
 
-    public Double getYear2SourceActiveLicense() {
+    public Integer getYear2SourceActiveLicense() {
         return year2SourceActiveLicense;
     }
 
-    public void setYear2SourceActiveLicense(Double year2SourceActiveLicense) {
+    public void setYear2SourceActiveLicense(Integer year2SourceActiveLicense) {
         this.year2SourceActiveLicense = year2SourceActiveLicense;
     }
 
-    public Double getYear2SourcePaidMaintenance() {
+    public Integer getYear2SourcePaidMaintenance() {
         return year2SourcePaidMaintenance;
     }
 
-    public void setYear2SourcePaidMaintenance(Double year2SourcePaidMaintenance) {
+    public void setYear2SourcePaidMaintenance(Integer year2SourcePaidMaintenance) {
         this.year2SourcePaidMaintenance = year2SourcePaidMaintenance;
     }
 
@@ -101,27 +101,27 @@ public class SourceRampDownCostsModel
         this.year2SourceMaintenanceTotalValue = year2SourceMaintenanceTotalValue;
     }
 
-    public Double getYear3ServersOffSource() {
+    public Integer getYear3ServersOffSource() {
         return year3ServersOffSource;
     }
 
-    public void setYear3ServersOffSource(Double year3ServersOffSource) {
+    public void setYear3ServersOffSource(Integer year3ServersOffSource) {
         this.year3ServersOffSource = year3ServersOffSource;
     }
 
-    public Double getYear3SourceActiveLicense() {
+    public Integer getYear3SourceActiveLicense() {
         return year3SourceActiveLicense;
     }
 
-    public void setYear3SourceActiveLicense(Double year3SourceActiveLicense) {
+    public void setYear3SourceActiveLicense(Integer year3SourceActiveLicense) {
         this.year3SourceActiveLicense = year3SourceActiveLicense;
     }
 
-    public Double getYear3SourcePaidMaintenance() {
+    public Integer getYear3SourcePaidMaintenance() {
         return year3SourcePaidMaintenance;
     }
 
-    public void setYear3SourcePaidMaintenance(Double year3SourcePaidMaintenance) {
+    public void setYear3SourcePaidMaintenance(Integer year3SourcePaidMaintenance) {
         this.year3SourcePaidMaintenance = year3SourcePaidMaintenance;
     }
